Fix for 2Gb limit on TDFWriter
git-svn-id: https://igv.googlecode.com/svn/trunk@420 17392f64-ead8-4cea-ae29-09b3ab513800

+++ src/org/broad/igv/tdf/TDFWriter.java
@@ -48,7 +48,7 @@ public class TDFWriter {
     static private Logger log = Logger.getLogger(TDFWriter.class);
     static private int version = 3;
     OutputStream fos = null;
-    int bytesWritten = 0;
+    long bytesWritten = 0;
 
     File file;
     Map<String, TDFGroup> groupCache = new LinkedHashMap();
